# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Spendwise API",
    "version": "1.0.0",
    "description": "API documentation for Spendwise Personal Finance Tracker"
  },
  "servers": [
    {
      "url": "https://spendwise-api-zrh5.onrender.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserRegister": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string", "example": "John Doe" },
          "email": { "type": "string", "format": "email", "example": "john@example.com" },
          "password": { "type": "string", "example": "strongpassword" }
        }
      },
      "UserLogin": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "john@example.com" },
          "password": { "type": "string", "example": "strongpassword" }
        }
      },
      "GoogleAuthRequest": {
        "type": "object",
        "required": ["tokenId"],
        "properties": {
          "tokenId": { "type": "string", "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6Ij..." }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." },
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "string", "example": "64d2f27e6c56d644ab123456" },
              "name": { "type": "string", "example": "John Doe" },
              "email": { "type": "string", "format": "email", "example": "john@example.com" }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Invalid Google token" },
          "errors": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserRegister" } } }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": {
            "description": "Email already in use or validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user and get JWT token",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserLogin" } } }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } }
          },
          "400": {
            "description": "Invalid credentials",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/auth/google": {
      "post": {
        "tags": ["Auth"],
        "summary": "Google OAuth login/signup",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GoogleAuthRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Google login/signup",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } }
          },
          "400": {
            "description": "Missing tokenId",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Invalid Google token",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  }
}